move Helpers.map_values out of Gtfs
