+++ test/helpers_test.exs
@@ -1,0 +1,5 @@
+defmodule HelpersTest do
+  use ExUnit.Case, async: true
+
+  doctest Helpers
+end
